--- src/main/java/com/basho/riak/client/core/PoolStateListener.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2013 Brian Roach <roach at basho dot com>.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.basho.riak.client.core;
-
-/**
- *
- * @author Brian Roach <roach at basho dot com>
- * @since 2.0
- */
-public interface PoolStateListener
-{
-    void poolStateChanged(ConnectionPool pool, ConnectionPool.State newState);
-}
