+++ src/main/java/com/basho/riak/client/core/RiakNode.java
@@ -15,16 +15,17 @@
  */
 package com.basho.riak.client.core;
 
+import com.basho.riak.client.core.netty.RiakChannelInitializer;
 import com.basho.riak.client.util.Constants;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelFutureListener;
-import io.netty.channel.ChannelInitializer;
+import io.netty.channel.ChannelOption;
 import io.netty.channel.nio.NioEventLoopGroup;
-import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioSocketChannel;
 import io.netty.handler.timeout.ReadTimeoutHandler;
+import java.net.InetSocketAddress;
 import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -34,8 +35,12 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.Executors;
+import java.util.concurrent.LinkedBlockingDeque;
 import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,7 +50,7 @@ import org.slf4j.LoggerFactory;
  * @author Brian Roach <roach at basho dot com>
  * @since 2.0
  */
-public class RiakNode implements ChannelFutureListener, RiakResponseListener, PoolStateListener
+public class RiakNode implements RiakResponseListener
 {
     public enum State
     {
@@ -53,19 +58,68 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
     }
     
     private final Logger logger = LoggerFactory.getLogger(RiakNode.class);
-    private final ConnectionPool connectionPool;
+    
+    private final LinkedBlockingDeque<ChannelWithIdleTime> available =
+        new LinkedBlockingDeque<ChannelWithIdleTime>();
+    private final ConcurrentLinkedQueue<Channel> inUse = 
+        new ConcurrentLinkedQueue<Channel>();
+    private final ConcurrentLinkedQueue<ChannelWithIdleTime> recentlyClosed =
+        new ConcurrentLinkedQueue<ChannelWithIdleTime>();
     private final List<NodeStateListener> stateListeners = 
         Collections.synchronizedList(new LinkedList<NodeStateListener>());
+    private final Map<Channel, FutureOperation> inProgressMap = 
+        new ConcurrentHashMap<Channel, FutureOperation>();
     
-    private volatile ScheduledExecutorService executor; 
-    private volatile boolean ownsExecutor;
+    private final Sync permits;
+    private final String remoteAddress;
+    private final int port;
     private volatile Bootstrap bootstrap;
     private volatile boolean ownsBootstrap;
+    private volatile ScheduledExecutorService executor;
+    private volatile boolean ownsExecutor;
     private volatile State state;
+    private volatile ScheduledFuture<?> idleReaperFuture;
+    private volatile ScheduledFuture<?> healthMonitorFuture;
+    private volatile int minConnections;
+    private volatile long idleTimeoutInNanos;
+    private volatile int connectionTimeout;
     
-    private Map<Channel, FutureOperation> inProgressMap = 
-        new ConcurrentHashMap<Channel, FutureOperation>();
+    
     private volatile int readTimeoutInMillis;
+    
+    private ChannelFutureListener writeListener =
+        new ChannelFutureListener() {
+
+        @Override
+        public void operationComplete(ChannelFuture future) throws Exception
+        {
+            // See how the write worked out ...
+            if (!future.isSuccess())
+            {
+                FutureOperation inProgress = inProgressMap.remove(future.channel());
+                logger.info("Write failed on RiakNode {}:{}", remoteAddress, port);
+                returnConnection(future.channel());
+                inProgress.setException(future.cause());
+            }
+        }
+            
+    };
+    
+    private final ChannelFutureListener closeListener = 
+        new ChannelFutureListener() {
+
+        @Override
+        public void operationComplete(ChannelFuture future) throws Exception
+        {
+            // Because we aren't storing raw channels in available we just throw
+            // these in the recentlyClosed as an indicator. We'll leave it up
+            // to the healthcheck task to purge the available queue of dead
+            // connections and make decisions on what to do 
+            recentlyClosed.add(new ChannelWithIdleTime(future.channel()));
+            logger.debug("Channel closed; id:{} {}:{}", future.channel().hashCode(), remoteAddress, port);
+        }
+     };
+    
     
     // TODO: Harden to prevent operation from being executed > 1 times?
     // TODO: how many channels on one event loop? 
@@ -73,13 +127,28 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
     {
         this.readTimeoutInMillis = builder.readTimeout;
         this.executor = builder.executor;
+        this.connectionTimeout = builder.connectionTimeout;
+        this.idleTimeoutInNanos = TimeUnit.NANOSECONDS.convert(builder.idleTimeout, TimeUnit.MILLISECONDS);
+        this.minConnections = builder.minConnections;
+        this.port = builder.port;
+        this.remoteAddress = builder.remoteAddress;
         
         if (builder.bootstrap != null)
         {
             this.bootstrap = builder.bootstrap.clone();
         }
         
-        this.connectionPool = builder.poolBuilder.build();
+        if (builder.maxConnections < 1)
+        {
+            permits = new Sync(Integer.MAX_VALUE);
+        }
+        else
+        {
+            permits = new Sync(builder.maxConnections);
+        }
+        
+        
+        
         this.state = State.CREATED;
     }
         
@@ -87,8 +156,8 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
     {
         if (Arrays.binarySearch(allowedStates, state) < 0)
         {
-            logger.debug("IllegalStateException; remote: {} required: {} current: {} ",
-                         connectionPool.getRemoteAddress(), Arrays.toString(allowedStates), state);
+            logger.debug("IllegalStateException; RiakNode: {}:{} required: {} current: {} ",
+                         remoteAddress, port, Arrays.toString(allowedStates), state);
             throw new IllegalStateException("required: " 
                 + Arrays.toString(allowedStates) 
                 + " current: " + state );
@@ -113,12 +182,42 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
             ownsBootstrap = true;
         }
         
-        connectionPool.addStateListener(this)
-                      .setBootstrap(bootstrap)
-                      .setExecutor(executor)
-                      .start();
+        bootstrap.handler(new RiakChannelInitializer(this))
+                 .remoteAddress(new InetSocketAddress(remoteAddress, port));
+
+        if (connectionTimeout > 0)
+        {
+            bootstrap.option(ChannelOption.CONNECT_TIMEOUT_MILLIS, connectionTimeout);
+        }
+        
+        if (minConnections > 0)
+        {
+            List<Channel> minChannels = new LinkedList<Channel>();
+            for (int i = 0; i < minConnections; i++)
+            {
+                Channel channel;
+                try
+                {
+                    channel = doGetConnection();
+                    minChannels.add(channel);
+                }
+                catch (ConnectionFailedException ex)
+                {
+                    // no-op, we don't care right now
+                }
+            }
+
+            for (Channel c : minChannels)
+            {
+                available.offerFirst(new ChannelWithIdleTime(c));
+            }
+        }
+
+        idleReaperFuture = executor.scheduleWithFixedDelay(new IdleReaper(), 1, 5, TimeUnit.SECONDS);
+        healthMonitorFuture = executor.scheduleWithFixedDelay(new HealthMonitorTask(), 1000, 500, TimeUnit.MILLISECONDS);
         
         state = State.RUNNING;
+        logger.info("RiakNode started; {}:{}", remoteAddress, port);
         notifyStateListeners();
         return this;
     }
@@ -126,21 +225,33 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
     public synchronized void shutdown()
     {
         stateCheck(State.RUNNING, State.HEALTH_CHECKING);
-        logger.info("Riak node shutting down; {}", connectionPool.getRemoteAddress());
-        connectionPool.shutdown();
-        // Notifications from the pool change our state. 
+        state = State.SHUTTING_DOWN;
+        logger.info("RiakNode shutting down; {}:{}", remoteAddress, port);
+        notifyStateListeners();
+        idleReaperFuture.cancel(true);
+        healthMonitorFuture.cancel(true);
+        ChannelWithIdleTime cwi = available.poll();
+        while (cwi != null)
+        {
+            Channel c = cwi.getChannel();
+            closeConnection(c);
+            cwi = available.poll();
+        }
+
+        executor.schedule(new ShutdownTask(), 0, TimeUnit.SECONDS);
     }
     
     /**
-     * Sets the Netty {@link Bootstrap} for this Node's connection pool(s).
+     * Sets the Netty {@link Bootstrap} for this Node's connections.
      * {@link Bootstrap#clone()} is called to clone the bootstrap.
      * 
-     * @param bootstrap
+     * @param bootstrap - the Netty Bootstrap to use
+     * @return a reference to this RiakNode
      * @throws IllegalArgumentException if it was already set via the builder.
      * @throws IllegalStateException if the node has already been started.
      * @see Builder#withBootstrap(io.netty.bootstrap.Bootstrap) 
      */
-    public void setBootstrap(Bootstrap bootstrap)
+    public RiakNode setBootstrap(Bootstrap bootstrap)
     {
         stateCheck(State.CREATED);
         if (this.bootstrap != null)
@@ -149,17 +260,19 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
         }
         
         this.bootstrap = bootstrap.clone();
+        return this;
     }
     
     /**
      * Sets the {@link ScheduledExecutorService} for this Node and its pool(s).
      * 
-     * @param executor
+     * @param executor - the ScheduledExecutorService to use.
+     * @return a reference to this RiakNode
      * @throws IllegalArgumentException if it was already set via the builder.
      * @throws IllegalStateException if the node has already been started.
      * @see Builder#withExecutor(java.util.concurrent.ScheduledExecutorService) 
      */
-    public void setExecutor(ScheduledExecutorService executor)
+    public RiakNode setExecutor(ScheduledExecutorService executor)
     {
         stateCheck(State.CREATED);
         if (this.executor != null)
@@ -167,6 +280,166 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
             throw new IllegalArgumentException("Executor already set");
         }
         this.executor = executor;
+        return this;
+    }
+    
+    /**
+     * Sets the read timeout for connections.
+     * @param readTimeoutInMillis the readTimeout to set
+     * @return a reference to this RiakNode
+     * @see Builder#withReadTimeout(int) 
+     */
+    public RiakNode setReadTimeout(int readTimeoutInMillis)
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        this.readTimeoutInMillis = readTimeoutInMillis;
+        return this;
+    }
+    
+    /**
+     * Returns the read timeout in milliseconds for connections.
+     * @return the readTimeout
+     * @see Builder#withReadTimeout(int) 
+     */
+    public int getReadTimeout()
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        return readTimeoutInMillis;
+    }
+
+    /**
+     * Sets the maximum number of connections allowed.
+     *
+     * @param maxConnections the maxConnections to set.
+     * @return a reference to this RiakNode.
+     * @see Builder#withMaxConnections(int) 
+     */
+    public RiakNode setMaxConnections(int maxConnections)
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        if (maxConnections >= getMinConnections())
+        {
+            permits.setMaxPermits(maxConnections);
+        }
+        else
+        {
+            throw new IllegalArgumentException("Max connections less than min connections");
+        }
+        // TODO: reap delta? 
+        return this;
+    }
+    
+    /**
+     * Returns the maximum number of connections allowed.
+     *
+     * @return the maxConnections
+     * @see Builder#withMaxConnections(int)
+     */
+    public int getMaxConnections()
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        return permits.getMaxPermits();
+    }
+    
+    /**
+     * Sets the minimum number of active connections to be maintained.
+     *
+     * @param minConnections the minConnections to set
+     * @return a reference to this RiakNode
+     * @see Builder#withMinConnections(int)
+     */
+    public RiakNode setMinConnections(int minConnections)
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        if (minConnections <= getMaxConnections())
+        {
+            this.minConnections = minConnections;
+        }
+        else
+        {
+            throw new IllegalArgumentException("Min connections greater than max connections");
+        }
+        // TODO: Start / reap delta?
+        return this;
+    }
+    
+    /**
+     * Returns the current minimum number of active connections to be maintained.
+     *
+     * @return the minConnections
+     * @see Builder#withMinConnections(int)
+     */
+    public int getMinConnections()
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        return minConnections;
+    }
+    
+    /**
+     * Sets the connection idle timeout for connections.
+     *
+     * @param idleTimeoutInMillis the idleTimeout to set
+     * @return a reference to this RiakNode
+     * @see Builder#withIdleTimeout(int)
+     */
+    public RiakNode setIdleTimeout(int idleTimeoutInMillis)
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        this.idleTimeoutInNanos = TimeUnit.NANOSECONDS.convert(idleTimeoutInMillis, TimeUnit.MILLISECONDS);
+        return this;
+    }
+    
+    /**
+     * Returns the connection idle timeout for connections in milliseconds.
+     *
+     * @return the idleTimeout in milliseconds
+     * @see Builder#withIdleTimeout(int)
+     */
+    public int getIdleTimeout()
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        return (int) TimeUnit.MILLISECONDS.convert(idleTimeoutInNanos, TimeUnit.NANOSECONDS);
+    }
+    
+    /**
+     * Sets the connection timeout for new connections. 
+     *
+     * @param connectionTimeoutInMillis the connectionTimeout to set
+     * @return a reference to this RiakNode
+     * @see Builder#withConnectionTimeout(int)
+     */
+    public RiakNode setConnectionTimeout(int connectionTimeoutInMillis)
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        this.connectionTimeout = connectionTimeoutInMillis;
+        bootstrap.option(ChannelOption.CONNECT_TIMEOUT_MILLIS, connectionTimeout);
+        return this;
+    }
+    
+    /**
+     * Returns the connection timeout in milliseconds.
+     *
+     * @return the connectionTimeout
+     * @see Builder#withConnectionTimeout(int)
+     */
+    public int getConnectionTimeout()
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        return connectionTimeout;
+    }
+    
+    /**
+     * Returns the number of permits currently available.
+     * The number of available permits indicates how many additional
+     * connections can be made without blocking.
+     *
+     * @return the number of available permits.
+     * @see Builder#withMaxConnections(int)
+     */
+    public int availablePermits()
+    {
+        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
+        return permits.availablePermits();
     }
     
     public void addStateListener(NodeStateListener listener)
@@ -206,7 +479,7 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
         stateCheck(State.RUNNING, State.HEALTH_CHECKING);
         
         operation.setLastNode(this);
-        Channel channel = connectionPool.getConnection();
+        Channel channel = getConnection();
         if (channel != null)
         {
             // Add a timeout handler to the pipeline if the readTIeout is set
@@ -217,12 +490,10 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
                              new ReadTimeoutHandler(readTimeoutInMillis, TimeUnit.MILLISECONDS));
             }
             
-            //TODO: figure out a cleaner way to do this? 
-            channel.pipeline().get(Constants.RESPONSE_HANDLER_CLASS).setListener(this);
             inProgressMap.put(channel, operation);
             ChannelFuture writeFuture = channel.writeAndFlush(operation); 
-            writeFuture.addListener(this);
-            logger.debug("Operation executed on node {} {}", connectionPool.getRemoteAddress(), channel.remoteAddress());
+            writeFuture.addListener(writeListener);
+            logger.debug("Operation executed on RiakNode {}:{}", remoteAddress, port);
             return true;
         }
         else
@@ -231,65 +502,145 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
         }
     }
     
-    @Override
-    public synchronized void poolStateChanged(ConnectionPool pool, ConnectionPool.State newState)
+    // ConnectionPool Stuff
+    
+    /**
+     * Get a Netty channel from the pool.
+     *
+     * @return a connected channel or {@code null} if all connections are in use or
+     *         a new connection can not be made.
+     */
+    private Channel getConnection()
     {
-        switch (newState)
+        stateCheck(State.RUNNING, State.HEALTH_CHECKING);
+        boolean acquired = permits.tryAcquire();
+        Channel channel = null;
+        if (acquired)
         {
-            case HEALTH_CHECKING:
-                this.state = State.HEALTH_CHECKING;
-                notifyStateListeners();
-                logger.info("RiakNode offline, health checking; {}", connectionPool.getRemoteAddress());
+            try
+            {
+                channel = doGetConnection();
+                inUse.add(channel);
+            }
+            catch (ConnectionFailedException ex)
+            {
+                permits.release();
+            }
+        }
+        return channel;
+    }
+
+    private Channel doGetConnection() throws ConnectionFailedException
+    {
+        ChannelWithIdleTime cwi;
+        while ((cwi = available.poll()) != null)
+        {
+            Channel channel = cwi.getChannel();
+            // If the channel from available is closed, try again. This will result in
+            // the caller always getting a connection or an exception. If closed
+            // the channel is simply discarded so this also acts as a purge
+            // for dead channels during a health check.
+            if (channel.isOpen())
+            {
+                return channel;
+            }
+        }
+
+
+        ChannelFuture f = bootstrap.connect();
+        // Any channels that don't connect will trigger a close operation as well
+        f.channel().closeFuture().addListener(closeListener);
+
+        try
+        {
+            f.await();
+        }
+        catch (InterruptedException ex)
+        {
+            logger.info("Thread interrupted waiting for new connection to be made; {}",
+                remoteAddress);
+            Thread.currentThread().interrupt();
+            throw new ConnectionFailedException(ex);
+        }
+
+        if (!f.isSuccess())
+        {
+            logger.error("Connection attempt failed: {}:{}; {}",
+                remoteAddress, port, f.cause());
+            throw new ConnectionFailedException(f.cause());
+        }
+
+        return f.channel();
+
+    }
+    
+    /**
+     * Return a Netty channel to the pool.
+     *
+     * @param c The Netty channel to return to the pool
+     * @throws IllegalArgumentException If the channel did not originate from this pool
+     */
+    private void returnConnection(Channel c)
+    {
+        stateCheck(State.RUNNING, State.SHUTTING_DOWN, State.SHUTDOWN, State.HEALTH_CHECKING);
+        if (!inUse.remove(c))
+        {
+            throw new IllegalArgumentException("Channel not managed by this pool");
+        }
+
+        switch (state)
+        {
+            case SHUTTING_DOWN:
+            case SHUTDOWN:
+                closeConnection(c);
                 break;
             case RUNNING:
-                this.state = State.RUNNING;
-                notifyStateListeners();
-                logger.info("RiakNode running; {}", connectionPool.getRemoteAddress());
-                break;
-            case SHUTTING_DOWN:
-                if (this.state == State.RUNNING ||  this.state == State.HEALTH_CHECKING)
-                {
-                    this.state = State.SHUTTING_DOWN;
-                    notifyStateListeners();
-                    logger.info("RiakNode shutting down due to pool shutdown; {}", connectionPool.getRemoteAddress());
-                }
-                break;
-            case SHUTDOWN:
-                this.state = State.SHUTDOWN;
-                notifyStateListeners();
-                if (ownsExecutor)
-                {
-                    executor.shutdown();
-                }
-                if (ownsBootstrap)
-                {
-                    bootstrap.group().shutdownGracefully();
-                }
-                logger.info("RiakNode shut down due to pool shutdown; {}", connectionPool.getRemoteAddress());
-                break;
+            case HEALTH_CHECKING:
             default:
+                if (c.isOpen())
+                {
+                    logger.debug("Channel id:{} returned to pool", c.hashCode());
+                    available.offerFirst(new ChannelWithIdleTime(c));
+                }
+                else
+                {
+                    logger.debug("Closed channel id:{} returned to pool; discarding", c.hashCode());
+                }
+                permits.release();
                 break;
         }
     }
-
+    
+    private void closeConnection(Channel c)
+    {
+        // If we are explicitly closing the connection we don't want to hear
+        // about it.
+        c.closeFuture().removeListener(closeListener);
+        c.close();
+    }
+    
+    
+    // End ConnectionPool stuff
+    
     @Override
     public void onSuccess(Channel channel, RiakMessage response)
     {
-        logger.debug("Operation onSuccess() channel: id:{} {}", channel.hashCode(), channel.remoteAddress());
+        logger.debug("Operation onSuccess() channel: id:{} {}:{}", channel.hashCode(), 
+                                                                   remoteAddress, port);
         if (readTimeoutInMillis > 0)
         {
             channel.pipeline().remove(Constants.TIMEOUT_HANDLER);
         }
         FutureOperation inProgress = inProgressMap.remove(channel);
-        connectionPool.returnConnection(channel);
+        returnConnection(channel);
         inProgress.setResponse(response);
     }
 
     @Override
     public void onException(Channel channel, Throwable t)
     {
-        logger.debug("Operation onException() channel: id:{} {} {}", 
-                     channel.hashCode(), channel.remoteAddress(), t);
+        logger.debug("Operation onException() channel: id:{} {}:{} {}", 
+                     channel.hashCode(), remoteAddress, port, t);
         
         FutureOperation inProgress = inProgressMap.remove(channel);
         // There are fail cases where multiple exceptions are thrown from 
@@ -302,26 +653,10 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
             {
                 channel.pipeline().remove(Constants.TIMEOUT_HANDLER);
             }
-            connectionPool.returnConnection(channel);
+            returnConnection(channel);
             inProgress.setException(t);
         }
     }
-    
-    // The only Netty future we are listening to is the WriteFuture
-    @Override
-    public void operationComplete(ChannelFuture future) throws Exception
-    {
-        // See how the write worked out ...
-        if (!future.isSuccess())
-        {
-            FutureOperation inProgress = inProgressMap.remove(future.channel());
-            logger.info("Write failed on node {}", connectionPool.getRemoteAddress());
-            connectionPool.returnConnection(future.channel());
-            inProgress.setException(future.cause());
-        }
-    }
-    
-    
     
     /**
      * Returns the {@code remoteAddress} for this RiakNode
@@ -329,7 +664,16 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
      */
     public String getRemoteAddress()
     {
-        return connectionPool.getRemoteAddress();
+        return remoteAddress;
+    }
+    
+    /**
+     * returns the remote port for this RiakNode
+     * @return the port number
+     */
+    public int getPort()
+    {
+        return port;
     }
     
     /**
@@ -341,36 +685,254 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
         return this.state;
     }
     
-    /**
-     * Returns the read timeout in milliseconds for connections in this pool
-     * @return the readTimeout
-     * @see Builder#withReadTimeout(int) 
-     */
-    public int getReadTimeout()
+    private class ChannelWithIdleTime
     {
-        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
-        return readTimeoutInMillis;
-    }
+        private Channel channel;
+        private long idleStart;
 
-    /**
-     * Sets the read timeout for connections in this pool
-     * @param readTimeoutInMillis the readTimeout to set
-     * @see Builder#withReadTimeout(int) 
-     */
-    public void setReadTimeout(int readTimeoutInMillis)
+        public ChannelWithIdleTime(Channel channel)
+        {
+            this.channel = channel;
+            idleStart = System.nanoTime();
+        }
+
+        public Channel getChannel()
+        {
+            return channel;
+        }
+
+        public long getIdleStart()
+        {
+            return idleStart;
+        }
+    }
+    
+    private class Sync extends Semaphore
     {
-        stateCheck(State.CREATED, State.RUNNING, State.HEALTH_CHECKING);
-        this.readTimeoutInMillis = readTimeoutInMillis;
+        private static final long serialVersionUID = -5118488872281021072L;
+        private volatile int maxPermits;
+
+        public Sync(int numPermits)
+        {
+            super(numPermits);
+            this.maxPermits = numPermits;
+        }
+
+        public Sync(int numPermits, boolean fair)
+        {
+            super(numPermits, fair);
+            this.maxPermits = numPermits;
+        }
+
+        public int getMaxPermits()
+        {
+            return maxPermits;
+        }
+
+        // Synchronized because we're (potentially) changing this.maxPermits
+        synchronized void setMaxPermits(int maxPermits)
+        {
+            int diff = maxPermits - this.maxPermits;
+
+            if (diff == 0)
+            {
+                return;
+            }
+            else if (diff > 0)
+            {
+                release(diff);
+            }
+            else if (diff < 0)
+            {
+                reducePermits(diff);
+            }
+
+            this.maxPermits = maxPermits;
+        }
+
+    }
+    
+    private class IdleReaper implements Runnable
+    {
+        @Override
+        public void run()
+        {
+            reapIdleConnections();
+        }
+    }
+    
+    private void reapIdleConnections()
+    {
+        // with all the concurrency there's really no reason to keep 
+        // checking the sizes. This is really just a "best guess"
+        int currentNum = inUse.size() + available.size();
+        if (currentNum > minConnections)
+        {
+            // Note this will not throw a ConncurrentModificationException
+            // and if hasNext() returns true you are guaranteed that
+            // the next() will return a value (even if it has already
+            // been removed from the Deque between those calls). 
+            Iterator<ChannelWithIdleTime> i = available.descendingIterator();
+            while (i.hasNext() && currentNum > minConnections)
+            {
+                ChannelWithIdleTime cwi = i.next();
+                if (cwi.getIdleStart() + idleTimeoutInNanos < System.nanoTime())
+                {
+                    boolean removed = available.remove(cwi);
+                    if (removed)
+                    {
+                        Channel c = cwi.getChannel();
+                        logger.debug("Idle channel closed; {}:{}", remoteAddress, port);
+                        closeConnection(c);
+                        currentNum--;
+                    }
+                }
+                else
+                {
+                    // Since we are descending and this is a LIFO, 
+                    // if the current connection hasn't been idle beyond 
+                    // the threshold, there's no reason to descend further
+                    break;
+                }
+            }
+        }
+    }
+    
+    // TODO: Magic numbers; probably should be configurable and less magic
+    private class HealthMonitorTask implements Runnable
+    {
+        @Override
+        public void run()
+        {
+            if ((state == State.RUNNING || state == State.HEALTH_CHECKING) &&
+                (recentlyClosed.size() > 4 || state == State.HEALTH_CHECKING))
+            {
+                checkHealth();
+            }
+        }
+    }
+    
+    // TODO: Magic numbers; probably should be configurable and less magic
+    // TODO: Maybe ping the node or something more conclusive than if it just accepts a connection
+    private void checkHealth()
+    {
+        try
+        {
+            // purge recentlyClosed past a certain age
+            // sliding window should be larger than the
+            // frequency of this task
+            long current = System.nanoTime();
+            long window = 3000000000L; // 3 seconds 
+            for (ChannelWithIdleTime cwi = recentlyClosed.peek();
+                 cwi != null && current - cwi.getIdleStart() > window;
+                 cwi = recentlyClosed.peek())
+            {
+                recentlyClosed.poll();
+            }
+
+            // See: doGetConnection() - this will purge closed
+            // connections from the available queue and either 
+            // return/create a new one (meaning the node is up) or throw
+            // an exception if a connection can't be made.
+            Channel c = doGetConnection();
+            closeConnection(c);
+
+            if (state == State.HEALTH_CHECKING)
+            {
+                logger.info("RiakNode recovered; {}:{}", remoteAddress, port);
+                state = State.RUNNING;
+                notifyStateListeners();
+            }
+
+        }
+        catch (ConnectionFailedException ex)
+        {
+            if (state == State.RUNNING)
+            {
+                logger.error("RiakNode offline; health checking; {}:{} {}",
+                    remoteAddress, port, ex);
+                state = State.HEALTH_CHECKING;
+                notifyStateListeners();
+            }
+            else
+            {
+                logger.error("RiakNode failed health check; {}:{} {}",
+                    remoteAddress, port, ex);
+            }
+        }
+        catch (IllegalStateException e)
+        {
+            // no-op; there's a race condition where the bootstrap is shutting down
+            // right when a healthcheck occurs and netty will throw this
+        }
+
+
+    }
+    
+    private class ShutdownTask implements Runnable
+    {
+        @Override
+        public void run()
+        {
+            if (inUse.isEmpty())
+            {
+                state = State.SHUTDOWN;
+                notifyStateListeners();
+                if (ownsExecutor)
+                {
+                    executor.shutdown();
+                }
+                if (ownsBootstrap)
+                {
+                    bootstrap.group().shutdownGracefully();
+                }
+                logger.debug("RiakNode shut down {}:{}", remoteAddress, port);
+            }
+        }
     }
     
     /**
      * Builder used to construct a RiakNode.
-     * Most of these options are passed directly to a {@link ConnectionPool.Builder}. 
-     * The default values are documented in that class. 
      * 
      */
     public static class Builder
     {
+        /**
+         * The default remote address to be used if not specified: {@value #DEFAULT_REMOTE_ADDRESS}
+         *
+         * @see #withRemoteAddress(java.lang.String)
+         */
+        public final static String DEFAULT_REMOTE_ADDRESS = "127.0.0.1";
+        /**
+         * The default port number to be used if not specified: {@value #DEFAULT_REMOTE_PORT}
+         * 
+         * @see AbstractCollection#withRemotePort(int)
+         */
+        public final static int DEFAULT_REMOTE_PORT = 8087;
+        /**
+         * The default minimum number of connections to maintain if not specified: {@value #DEFAULT_MIN_CONNECTIONS}
+         *
+         * @see #withMinConnections(int)
+         */
+        public final static int DEFAULT_MIN_CONNECTIONS = 1;
+        /**
+         * The default maximum number of connections allowed if not specified: {@value #DEFAULT_MAX_CONNECTIONS}
+         *
+         * @see #withMaxConnections(int)
+         */
+        public final static int DEFAULT_MAX_CONNECTIONS = 0;
+        /**
+         * The default idle timeout in milliseconds for connections if not specified: {@value #DEFAULT_IDLE_TIMEOUT}
+         *
+         * @see #withIdleTimeout(int)
+         */
+        public final static int DEFAULT_IDLE_TIMEOUT = 1000;
+        /**
+         * The default connection timeout in milliseconds if not specified: {@value #DEFAULT_CONNECTION_TIMEOUT}
+         *
+         * @see #withConnectionTimeout(int)
+         */
+        public final static int DEFAULT_CONNECTION_TIMEOUT = 0;
         /**
          * The default TCP read timeout in milliseconds if not specified: {@value #DEFAULT_TCP_READ_TIMEOUT}
          * A value of {@code 0} means to wait indefinitely 
@@ -378,11 +940,18 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
          */
         public final static int DEFAULT_TCP_READ_TIMEOUT = 0;
         
-        private final ConnectionPool.Builder poolBuilder = new ConnectionPool.Builder();
-        
-        private ScheduledExecutorService executor;
-        private Bootstrap bootstrap;
+        private int port = DEFAULT_REMOTE_PORT;
+        private String remoteAddress = DEFAULT_REMOTE_ADDRESS;
+        private int minConnections = DEFAULT_MIN_CONNECTIONS;
+        private int maxConnections = DEFAULT_MAX_CONNECTIONS;
+        private int idleTimeout = DEFAULT_IDLE_TIMEOUT;
+        private int connectionTimeout = DEFAULT_CONNECTION_TIMEOUT;
         private int readTimeout = DEFAULT_TCP_READ_TIMEOUT;
+        private Bootstrap bootstrap;
+        private ScheduledExecutorService executor;
+        
+        
+        
         
         /**
          * Default constructor. Returns a new builder for a RiakNode with 
@@ -397,81 +966,99 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
          * Sets the remote address for this RiakNode. 
          * @param remoteAddress Can either be a FQDN or IP address
          * @return this
+         * @see #DEFAULT_REMOTE_ADDRESS
          */
         public Builder withRemoteAddress(String remoteAddress)
         {
-            poolBuilder.withRemoteAddress(remoteAddress);
+            this.remoteAddress = remoteAddress;
             return this;
         }
         
        /**
          * Specifies the remote port for this RiakNode.
-         * @param p - the protocol
          * @param port - the port
          * @return this
+         * @see #DEFAULT_REMOTE_PORT
          */
         public Builder withRemotePort(int port)
         {
-            poolBuilder.withRemotePort(port);
+            this.port = port;
             return this;
         }
         
         /**
-         * Specifies the minimum number for connections to be maintained
-         * @param p 
-         * @param minConnections
+         * Set the minimum number of active connections to maintain.
+         * These connections are exempt from the idle timeout.
+         * @param minConnections - number of connections to maintain.
          * @return this
-         * @see ConnectionPool.Builder#withMinConnections(int) 
+         * @see #DEFAULT_MIN_CONNECTIONS
          */
         public Builder withMinConnections(int minConnections)
         {
-            poolBuilder.withMinConnections(minConnections);
+            if (maxConnections == DEFAULT_MAX_CONNECTIONS || minConnections <= maxConnections)
+            {
+                this.minConnections = minConnections;
+            }
+            else
+            {
+                throw new IllegalArgumentException("Min connections greater than max connections");
+            }
             return this;
         }
         
         /**
-         * Specifies the maximum number of connections allowed.
-         * @param p
-         * @param maxConnections
+         * Set the maximum number of connections allowed. 
+         * A value of 0 sets this to unlimited.
+         * @param maxConnections - maximum number of connections to allow
          * @return this
-         * @see ConnectionPool.Builder#withMaxConnections(int) 
+         * @see #DEFAULT_MAX_CONNECTIONS
          */
         public Builder withMaxConnections(int maxConnections)
         {
-            poolBuilder.withMaxConnections(maxConnections);
+            if (maxConnections >= minConnections)
+            {
+                this.maxConnections = maxConnections;
+            }
+            else
+            {
+                throw new IllegalArgumentException("Max connections less than min connections");
+            }
             return this;
         }
         
         /**
-         * Specifies the idle timeout for connections.
-         * @param p
-         * @param idleTimeoutInMillis
+         * Set the idle timeout used to reap inactive connections.
+         * Any connection that has been idle for this amount of time
+         * becomes eligible to be closed and discarded unless {@code minConnections}
+         * has been set via {@link #withMinConnections(int) }
+         *
+         * @param idleTimeoutInMillis - idle timeout in milliseconds
          * @return this
-         * @see ConnectionPool.Builder#withIdleTimeout(int) 
+         * @see #DEFAULT_IDLE_TIMEOUT
          */
         public Builder withIdleTimeout(int idleTimeoutInMillis)
         {
-            poolBuilder.withIdleTimeout(idleTimeoutInMillis);
+            this.idleTimeout = idleTimeoutInMillis;
             return this;
         }
         
-        /**
-         * Specifies the connection timeout when creating new connections.
-         * @param p
+       /**
+         * Set the connection timeout used when making new connections
+         *
          * @param connectionTimeoutInMillis
          * @return this
-         * @see ConnectionPool.Builder#withConnectionTimeout(int) 
+         * @see #DEFAULT_CONNECTION_TIMEOUT
          */
         public Builder withConnectionTimeout(int connectionTimeoutInMillis)
         {
-            poolBuilder.withConnectionTimeout(connectionTimeoutInMillis);
+            this.connectionTimeout = connectionTimeoutInMillis;
             return this;
         }
         
         //TODO: Now that we have operation timeouts, do we really want to expose the TCP read timeout?
         /**
          * Specifies the TCP read timeout when waiting for a reply from Riak.
-         * @param readTimeoutInMillis
+         * @param readTimeoutInMillis - a timeout in milliseconds
          * @return this
          * @see #DEFAULT_READ_TIMEOUT
          */
@@ -483,10 +1070,10 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
         
         /**
          * Provides an executor for this node to use for internal maintenance tasks.
-         * This same executor will be used for this node's connection pool
-         * @param executor
+         * If not provided one will be created via
+         * {@link Executors#newSingleThreadScheduledExecutor()}
+         * @param executor the ScheduledExecutorService to use.
          * @return this
-         * @see ConnectionPool.Builder#withExecutor(java.util.concurrent.ScheduledExecutorService) 
          */
         public Builder withExecutor(ScheduledExecutorService executor)
         {
@@ -496,29 +1083,16 @@ public class RiakNode implements ChannelFutureListener, RiakResponseListener, Po
         
         /**
          * Provides a Netty Bootstrap for this node to use. 
-         * This same Bootstrap will be used for this node's underlying connection pool(s)
+         * If not provided one
+         * will be created with its own {@code NioEventLoopGroup}.
          * @param bootstrap
          * @return this
-         * @see ConnectionPool.Builder#withBootstrap(io.netty.bootstrap.Bootstrap) 
          */
         public Builder withBootstrap(Bootstrap bootstrap)
         {
             this.bootstrap = bootstrap;
             return this;
         }
-        
-        /**
-         * The Netty {@code ChannelInitializer} to use with the connection pool.
-         * @param initializer the initializer
-         * @return this
-         * @see ConnectionPool.Builder#withChannelInitializer(io.netty.channel.ChannelInitializer) 
-         */
-        public Builder withChannelInitializer(ChannelInitializer<SocketChannel> initializer)
-        {
-            poolBuilder.withChannelInitializer(initializer);
-            return this;
-        }
-        
         
         
         /**
